restructure hapi tests
